refactor: split instrumentation-build into per-concept modules
Move event-enum, record-struct, type-mapping, and shared codegen helpers
into events/records/data_type/common modules, mirroring the schema crate's
layout. Keep the public surface to GenerateOptions, GenerateError, generate,
and generate_str; per-module generators are crate-internal and tested by
unit tests in their own modules. Deconvolute the integration tests from
derive concerns: structural tests run with default (no-derive) options and
derive behaviour is covered by a dedicated cluster.
