+++ crates/instrumentation-build/src/lib.rs
@@ -21,8 +21,8 @@
 //!
 //! let schema = todo!(); // some schema (ideally validated, see `quent-constraints`)
 //! let opts = GenerateOptions {
-//!     event_derives: &["Debug", "Clone", "::serde::Serialize"],
-//!     record_derives: &["Debug", "Clone", "::serde::Serialize"],
+//!     event_derives: &["Debug", "Clone"],
+//!     record_derives: &["Debug", "Clone"],
 //!     out_dir: std::env::var("OUT_DIR")?.into(),
 //!     file_name: None, // defaults to `<schema name>.rs`
 //! };
@@ -37,13 +37,18 @@
 //! }
 //! ```
 
+mod common;
+mod data_type;
+mod events;
+mod records;
+
 use std::path::PathBuf;
 
-use convert_case::{Boundary, Case, Casing};
-use proc_macro2::{Span, TokenStream};
-use quent_schema::{DataType, Entity, Identifier, Record, Schema};
+use quent_schema::Schema;
 use quote::quote;
-use syn::Ident;
+
+use events::generate_event_types;
+use records::generate_record_types;
 
 /// Options controlling code generation.
 // TODO(johanpel): kept as simple as possible for now, but eventually some
@@ -87,13 +92,6 @@ pub enum GenerateError {
     Io(#[from] std::io::Error),
 }
 
-/// Combined token output: record structs followed by event enums.
-fn combined_tokens(schema: &Schema, opts: &GenerateOptions) -> Result<TokenStream, GenerateError> {
-    let records = generate_record_types(schema, opts)?;
-    let events = generate_event_types(schema, opts)?;
-    Ok(quote! { #records #events })
-}
-
 /// Generate the full instrumentation source for `schema` and write it to
 /// `opts.out_dir`, returning the path written.
 ///
@@ -104,10 +102,6 @@ fn combined_tokens(schema: &Schema, opts: &GenerateOptions) -> Result<TokenStrea
 ///
 /// Returns [`GenerateError`] if a derive entry is not a parseable Rust path, if
 /// the generated code is not a valid Rust file, or if writing the file fails.
-///
-/// # Panics
-///
-/// Panics if a field type nests deeper than [`MAX_TYPE_DEPTH`].
 pub fn generate(schema: &Schema, opts: &GenerateOptions) -> Result<PathBuf, GenerateError> {
     let file_name = opts
         .file_name
@@ -127,250 +121,11 @@ pub fn generate(schema: &Schema, opts: &GenerateOptions) -> Result<PathBuf, Gene
 ///
 /// Returns [`GenerateError`] if a derive entry is not a parseable Rust path, or
 /// if the generated code is not a valid Rust file.
-///
-/// # Panics
-///
-/// Panics if a field type nests deeper than [`MAX_TYPE_DEPTH`].
 pub fn generate_str(schema: &Schema, opts: &GenerateOptions) -> Result<String, GenerateError> {
-    let file = syn::parse2::<syn::File>(combined_tokens(schema, opts)?)
+    // record structs first, then event enums
+    let records = generate_record_types(schema, opts)?;
+    let events = generate_event_types(schema, opts)?;
+    let file = syn::parse2::<syn::File>(quote! { #records #events })
         .map_err(GenerateError::InvalidGeneratedCode)?;
     Ok(prettyplease::unparse(&file))
-}
-
-/// Per-entity event payload enums, as tokens.
-fn generate_event_types(
-    schema: &Schema,
-    opts: &GenerateOptions,
-) -> Result<TokenStream, GenerateError> {
-    let enums: Vec<TokenStream> = schema
-        .entities()
-        .map(|entity| entity_event_enum(entity, opts))
-        .collect::<Result<_, _>>()?;
-    Ok(quote! { #(#enums)* })
-}
-
-/// Generate the per-entity event payload enums for `schema`, in declaration
-/// order, formatted with `prettyplease`.
-///
-/// Each entity yields `pub enum <Entity>Event`; each of its events is a variant
-/// (UpperCamel) whose payload fields are snake_case named fields. Type, variant
-/// and field names are derived by case conversion that preserves digits, and
-/// names that are Rust keywords are raw-escaped. The caller must ensure entity,
-/// event and field names are unique after that conversion: two names that
-/// converge produce a duplicate definition that fails to compile.
-///
-/// # Errors
-///
-/// Returns [`GenerateError`] if a derive entry is not a parseable Rust path, or
-/// if the generated code is not a valid Rust file.
-///
-/// # Panics
-///
-/// Panics if a field type nests deeper than [`MAX_TYPE_DEPTH`].
-pub fn generate_event_types_str(
-    schema: &Schema,
-    opts: &GenerateOptions,
-) -> Result<String, GenerateError> {
-    let file = syn::parse2::<syn::File>(generate_event_types(schema, opts)?)
-        .map_err(GenerateError::InvalidGeneratedCode)?;
-    Ok(prettyplease::unparse(&file))
-}
-
-/// Record structs, as tokens.
-fn generate_record_types(
-    schema: &Schema,
-    opts: &GenerateOptions,
-) -> Result<TokenStream, GenerateError> {
-    let records: Vec<TokenStream> = schema
-        .records()
-        .map(|record| record_struct(record, opts))
-        .collect::<Result<_, _>>()?;
-    Ok(quote! { #(#records)* })
-}
-
-/// Generate the record structs for `schema`, in declaration order, formatted
-/// with `prettyplease`.
-///
-/// Each record yields `pub struct <Record>` with one public field per record
-/// field. Naming, raw-escaping and the uniqueness obligation match
-/// [`generate_event_types_str`].
-///
-/// # Errors
-///
-/// Returns [`GenerateError`] if a derive entry is not a parseable Rust path, or
-/// if the generated code is not a valid Rust file.
-///
-/// # Panics
-///
-/// Panics if a field type nests deeper than [`MAX_TYPE_DEPTH`].
-pub fn generate_record_types_str(
-    schema: &Schema,
-    opts: &GenerateOptions,
-) -> Result<String, GenerateError> {
-    let file = syn::parse2::<syn::File>(generate_record_types(schema, opts)?)
-        .map_err(GenerateError::InvalidGeneratedCode)?;
-    Ok(prettyplease::unparse(&file))
-}
-
-fn entity_event_enum(entity: &Entity, opts: &GenerateOptions) -> Result<TokenStream, GenerateError> {
-    let enum_ident = raw_ident(format!("{}Event", to_case(entity.name(), Case::Pascal)));
-    let docs = doc_attr(entity.annotations().docs());
-    let derives = derive_attr(opts.event_derives)?;
-    let variants: Vec<TokenStream> = entity
-        .events()
-        .map(|event| {
-            let variant = raw_ident(to_case(event.name(), Case::Pascal));
-            let variant_docs = doc_attr(event.annotations().docs());
-            let fields: Vec<TokenStream> = event
-                .fields()
-                .map(|field| {
-                    let name = raw_ident(to_case(field.name(), Case::Snake));
-                    let ty = map_data_type(field.ty());
-                    let field_docs = doc_attr(field.annotations().docs());
-                    quote! { #field_docs #name: #ty }
-                })
-                .collect();
-            if fields.is_empty() {
-                quote! { #variant_docs #variant }
-            } else {
-                quote! { #variant_docs #variant { #(#fields),* } }
-            }
-        })
-        .collect();
-    Ok(quote! {
-        #docs
-        #derives
-        pub enum #enum_ident {
-            #(#variants),*
-        }
-    })
-}
-
-fn record_struct(record: &Record, opts: &GenerateOptions) -> Result<TokenStream, GenerateError> {
-    let ident = raw_ident(to_case(record.name(), Case::Pascal));
-    let docs = doc_attr(record.annotations().docs());
-    let derives = derive_attr(opts.record_derives)?;
-    let fields: Vec<TokenStream> = record
-        .fields()
-        .map(|field| {
-            let name = raw_ident(to_case(field.name(), Case::Snake));
-            let ty = map_data_type(field.ty());
-            let field_docs = doc_attr(field.annotations().docs());
-            quote! { #field_docs pub #name: #ty }
-        })
-        .collect();
-    if fields.is_empty() {
-        Ok(quote! { #docs #derives pub struct #ident {} })
-    } else {
-        Ok(quote! {
-            #docs
-            #derives
-            pub struct #ident {
-                #(#fields),*
-            }
-        })
-    }
-}
-
-/// Maximum nesting depth of `Option`/`List`/`EntityRef` wrappers a single field
-/// type may have. A defensive bound against runaway recursion on malformed input;
-/// far above any realistic schema.
-pub const MAX_TYPE_DEPTH: usize = 64;
-
-/// Map a [`DataType`] to its Rust type tokens, recursing through `Option`,
-/// `List` and `EntityRef` payloads.
-fn map_data_type(ty: &DataType) -> TokenStream {
-    map_data_type_at(ty, 0)
-}
-
-fn map_data_type_at(ty: &DataType, depth: usize) -> TokenStream {
-    assert!(
-        depth <= MAX_TYPE_DEPTH,
-        "field type nesting exceeds the maximum depth of {MAX_TYPE_DEPTH}"
-    );
-    match ty {
-        DataType::Bool => quote! { bool },
-        DataType::Uuid => quote! { ::uuid::Uuid },
-        DataType::String => quote! { String },
-        DataType::U8 => quote! { u8 },
-        DataType::U16 => quote! { u16 },
-        DataType::U32 => quote! { u32 },
-        DataType::U64 => quote! { u64 },
-        DataType::I8 => quote! { i8 },
-        DataType::I16 => quote! { i16 },
-        DataType::I32 => quote! { i32 },
-        DataType::I64 => quote! { i64 },
-        DataType::F32 => quote! { f32 },
-        DataType::F64 => quote! { f64 },
-        DataType::Option(inner) => {
-            let inner = map_data_type_at(inner, depth + 1);
-            quote! { Option<#inner> }
-        }
-        DataType::List(inner) => {
-            let inner = map_data_type_at(inner, depth + 1);
-            quote! { Vec<#inner> }
-        }
-        DataType::Record(name) => {
-            let ident = raw_ident(to_case(name, Case::Pascal));
-            quote! { #ident }
-        }
-        DataType::DynamicRecord => quote! { ::quent_attributes::CustomAttributes },
-        DataType::EntityRef { data, .. } => match data {
-            Some(inner) => {
-                let inner = map_data_type_at(inner, depth + 1);
-                quote! { ::quent_instrumentation_runtime::EntityRef<#inner> }
-            }
-            None => quote! { ::quent_instrumentation_runtime::EntityRef },
-        },
-    }
-}
-
-fn derive_attr(derives: &[&str]) -> Result<TokenStream, GenerateError> {
-    if derives.is_empty() {
-        return Ok(quote! {});
-    }
-    let paths = derives
-        .iter()
-        .copied()
-        .map(|d| {
-            syn::parse_str::<syn::Path>(d).map_err(|source| GenerateError::InvalidDerive {
-                derive: d.to_owned(),
-                source,
-            })
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(quote! { #[derive(#(#paths),*)] })
-}
-
-fn doc_attr(docs: Option<&str>) -> TokenStream {
-    match docs {
-        Some(text) => quote! { #[doc = #text] },
-        None => quote! {},
-    }
-}
-
-/// Case-convert a schema identifier without splitting letter/digit boundaries,
-/// so names such as `u8` or `http2` are preserved rather than mangled.
-fn to_case(id: &Identifier, case: Case) -> String {
-    const KEEP_DIGITS: &[Boundary] = &[
-        Boundary::LOWER_DIGIT,
-        Boundary::UPPER_DIGIT,
-        Boundary::DIGIT_LOWER,
-        Boundary::DIGIT_UPPER,
-    ];
-    id.to_string().without_boundaries(KEEP_DIGITS).to_case(case)
-}
-
-/// Build an identifier from an already-cased name, raw-escaping Rust keywords.
-/// The keywords that cannot be raw (`crate`, `self`, `super`, `Self`) instead
-/// receive a trailing underscore.
-fn raw_ident(name: String) -> Ident {
-    const NON_RAW: &[&str] = &["crate", "self", "super", "Self"];
-    if NON_RAW.contains(&name.as_str()) {
-        Ident::new(&format!("{name}_"), Span::call_site())
-    } else if syn::parse_str::<Ident>(&name).is_ok() {
-        Ident::new(&name, Span::call_site())
-    } else {
-        Ident::new_raw(&name, Span::call_site())
-    }
 }
